securised all routes, update link between collections, fallback on user.remove

+++ src/routes/tasks_route.js
@@ -5,10 +5,30 @@ const auth = require("../middlewares/authentification");
 //Plural
 app
   .route("/tasks")
-  .get(async (req, res) => {
+  .get(auth, async (req, res) => {
+    const { completed, sortBy } = req.query;
+    const match = {};
+    const sort = {};
+
+    if (completed) match.completed = completed === "true" ? true : false;
+    if (sortBy) {
+      const parts = sortBy.split(":");
+      sort[parts[0]] = parts[1] === "asc" ? 1 : -1;
+    }
+
     try {
-      const tasks = await Task.find();
-      res.send(tasks);
+      await req.user
+        .populate({
+          path: "tasks",
+          match,
+          options: {
+            limit: parseInt(req.query.limit),
+            skip: parseInt(req.query.skip),
+            sort,
+          },
+        })
+        .execPopulate();
+      res.send(req.user.tasks);
     } catch (err) {
       res.status(400).send(err.message);
     }
@@ -25,9 +45,9 @@ app
       res.status(400).send(error.message);
     }
   })
-  .delete(async (req, res) => {
+  .delete(auth, async (req, res) => {
     try {
-      await Task.deleteMany();
+      await Task.deleteMany({ owner: req.user._id });
       res.send();
     } catch (error) {
       res.status(500).send();
@@ -38,9 +58,13 @@ app
 
 app
   .route("/tasks/:id")
-  .get(async (req, res) => {
+  .get(auth, async (req, res) => {
     try {
-      const task = await Task.findById(req.params.id);
+      const task = await Task.findOne({
+        _id: req.params.id,
+        owner: req.user._id,
+      });
+
       if (task) {
         res.send(task);
       } else {
@@ -50,26 +74,28 @@ app
       res.status(400).send(error.message);
     }
   })
-  .delete(async (req, res) => {
+  .delete(auth, async (req, res) => {
     try {
-      const result = await Task.findByIdAndRemove(req.params.id);
-      if (!result) return res.status(400).send("Invalid data");
-
+      const result = await Task.findOneAndRemove({
+        _id: req.params.id,
+        owner: req.user._id,
+      });
+      if (!result) return res.status(404).send();
       res.send(result);
     } catch (error) {
       res.status(500).send(error.message);
     }
   })
-  .patch(async (req, res) => {
+  .patch(auth, async (req, res) => {
     const { description, completed } = req.body;
-
-    if (!description || !completed) return res.status(400).send("Missing Data");
-
+    if (!description && !completed) return res.status(400).send();
     try {
-      const task = await Task.findById(req.params.id);
+      const task = await Task.findOne({
+        _id: req.params.id,
+        owner: req.user._id,
+      });
 
-      if (!task) return res.status(400).send("Didnt find the task");
-
+      if (!task) return res.status(404).send();
       Object.keys(req.body).forEach((key) => {
         task[key] = req.body[key];
       });
